requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
llama-cpp-python
langchain
fastapi
uvicorn
